move stub model function to test helper
