+++ spec/requests/provider_interface/offers_controller_spec.rb
@@ -8,6 +8,7 @@ end
 
 RSpec.describe ProviderInterface::OffersController, type: :request do
   include DfESignInHelpers
+  include ModelWithErrorsStubHelper
 
   let(:provider_user) { create(:provider_user, :with_dfe_sign_in, :with_make_decisions) }
   let(:provider) { provider_user.providers.first }
